+++ app/src/main/java/com/example/playlistmaker/media/domain/model/Playlist.kt
@@ -5,6 +5,6 @@ data class Playlist(
     val name: String,
     val description: String = "",
     val coverPath: String? = null,
-    val trackIds: List<Int> = emptyList(),
+    val trackIds: List<Long> = emptyList(),
     val tracksCount: Int = 0
 )
